# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Datapass Studio Root",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/exercises/{exercise_id}/solution": {
      "post": {
        "summary": "Reveal Solution",
        "operationId": "reveal_solution_api_exercises__exercise_id__solution_post",
        "parameters": [
          {
            "name": "exercise_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Exercise Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{id}/attempts": {
      "get": {
        "summary": "Attempts",
        "operationId": "attempts_api_workspaces__id__attempts_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExerciseAttempt"
                  },
                  "title": "Response Attempts Api Workspaces  Id  Attempts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{id}/practice/{exercise_id}/review": {
      "put": {
        "summary": "Review",
        "operationId": "review_api_workspaces__id__practice__exercise_id__review_put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "exercise_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Exercise Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{id}/exercise": {
      "post": {
        "summary": "Exercise Action",
        "operationId": "exercise_action_api_workspaces__id__exercise_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/workspaces": {
      "get": {
        "summary": "Workspaces",
        "operationId": "workspaces_api_workspaces_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "summary": "New Workspace",
        "operationId": "new_workspace_api_workspaces_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspace"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{id}": {
      "get": {
        "summary": "Workspace",
        "operationId": "workspace_api_workspaces__id__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{id}/notebook": {
      "put": {
        "summary": "Save",
        "operationId": "save_api_workspaces__id__notebook_put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveNotebook"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{id}/capabilities": {
      "get": {
        "summary": "Capabilities",
        "operationId": "capabilities_api_workspaces__id__capabilities_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{id}/catalog": {
      "get": {
        "summary": "Catalog",
        "operationId": "catalog_api_workspaces__id__catalog_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{id}/execute": {
      "post": {
        "summary": "Execute",
        "operationId": "execute_api_workspaces__id__execute_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteCell"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{id}/workflow": {
      "post": {
        "summary": "Workflow",
        "operationId": "workflow_api_workspaces__id__workflow_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunWorkflow"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{id}/check": {
      "post": {
        "summary": "Check",
        "operationId": "check_api_workspaces__id__check_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{id}/restart": {
      "post": {
        "summary": "Restart",
        "operationId": "restart_api_workspaces__id__restart_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateWorkspace": {
        "properties": {
          "case_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "CreateWorkspace"
      },
...
      "ExerciseAttempt": {
        "properties": {
          "schema_version": {
            "type": "integer",
            "const": 1,
            "title": "Schema Version"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "exercise_id": {
            "type": "string",
            "title": "Exercise Id"
          },
          "exercise_version": {
            "type": "string",
            "title": "Exercise Version"
          },
          "notebook_id": {
            "type": "string",
            "title": "Notebook Id"
          },
          "cell_id": {
            "type": "string",
            "title": "Cell Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "source_hash": {
            "type": "string",
            "title": "Source Hash"
          },
          "source_revision": {
            "type": "integer",
            "title": "Source Revision"
          },
          "validator_version": {
            "type": "string",
            "title": "Validator Version"
          },
          "fixtures": {
            "items": {
              "$ref": "#/components/schemas/VersionRef"
            },
            "type": "array",
            "title": "Fixtures"
          },
          "language": {
            "type": "string",
            "enum": [
              "sql",
              "sparklab",
              "python",
              "polars",
              "dbt"
            ],
            "title": "Language"
          },
          "status": {
            "type": "string",
            "enum": [
              "passed",
              "failed",
              "error"
            ],
            "title": "Status"
          },
          "checks": {
            "items": {
              "$ref": "#/components/schemas/ExerciseCheck"
            },
            "type": "array",
            "title": "Checks"
          },
          "truth": {
            "type": "string",
            "enum": [
              "real",
              "semantic-emulation",
              "simulated",
              "unsupported"
            ],
            "title": "Truth"
          },
          "runtime": {
            "$ref": "#/components/schemas/RuntimeIdentity"
          },
          "elapsed_ms": {
            "type": "number",
            "title": "Elapsed Ms"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExecutionError"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "schema_version",
          "id",
          "created_at",
          "exercise_id",
          "exercise_version",
          "notebook_id",
          "cell_id",
          "source",
          "source_hash",
          "source_revision",
          "validator_version",
          "fixtures",
          "language",
          "status",
          "checks",
          "truth",
          "runtime",
          "elapsed_ms"
        ],
        "title": "ExerciseAttempt"
      },
...
      "ExerciseRequest": {
        "properties": {
          "exercise_id": {
            "type": "string",
            "maxLength": 100,
            "title": "Exercise Id"
          },
          "exercise_version": {
            "type": "string",
            "maxLength": 40,
            "title": "Exercise Version"
          },
          "notebook_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Notebook Id"
          },
          "cell_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Cell Id"
          },
          "code": {
            "type": "string",
            "maxLength": 39000,
            "minLength": 1,
            "title": "Code"
          },
          "language": {
            "type": "string",
            "enum": [
              "sql",
              "sparklab",
              "python",
              "polars",
              "dbt"
            ],
            "title": "Language"
          },
          "source_revision": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Source Revision"
          },
          "mode": {
            "type": "string",
            "enum": [
              "run",
              "submit"
            ],
            "title": "Mode"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "exercise_id",
          "exercise_version",
          "notebook_id",
          "cell_id",
          "code",
          "language",
          "source_revision",
          "mode"
        ],
        "title": "ExerciseRequest"
      },
      "ExerciseResult": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "passed",
              "failed",
              "error"
            ],
            "title": "Status"
          },
          "checks": {
            "items": {
              "$ref": "#/components/schemas/ExerciseCheck"
            },
            "type": "array",
            "title": "Checks"
          },
          "runs": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Runs"
          },
          "attempt": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExerciseAttempt"
              },
              {
                "type": "null"
              }
            ]
          },
          "workspace_revision": {
            "type": "integer",
            "title": "Workspace Revision"
          },
          "truth": {
            "type": "string",
            "enum": [
              "real",
              "semantic-emulation",
              "simulated",
              "unsupported"
            ],
            "title": "Truth"
          },
          "runtime": {
            "$ref": "#/components/schemas/RuntimeIdentity"
          },
          "elapsed_ms": {
            "type": "number",
            "title": "Elapsed Ms"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExecutionError"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "checks",
          "runs",
          "workspace_revision",
          "truth",
          "runtime",
          "elapsed_ms"
        ],
        "title": "ExerciseResult"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReviewRequest": {
        "properties": {
          "revision": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Revision"
          },
          "review": {
            "type": "boolean",
            "title": "Review"
          },
          "confidence": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Confidence",
            "default": "low"
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "easy",
              "medium",
              "hard"
            ],
            "title": "Difficulty",
            "default": "easy"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "revision",
          "review"
        ],
        "title": "ReviewRequest"
      },
...
      "VersionRef": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "version": {
            "type": "string",
            "title": "Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "version"
        ],
        "title": "VersionRef"
      },
...
    }
  }
}